ROCK-8640 Backport the shared S&S connect gate and server-side enforcement from v16
Mirrors ROCK-8867 (PR #5, hotfix-1.16.12) onto 1.13.7 so both branches run the
same gate code:

- Adds RockWeb/App_Code/SeccConnectGateHelper.cs, byte-identical to the
  hotfix-1.16.12 copy. All gate logic (SecurityToConnect key variants, S&S role
  + Rock Administrators bypass, ConnectableStatuses) lives there.
- Both blocks' private gate implementations are replaced with thin wrappers
  that call the helper, removing the Board/Detail duplication PR #4 introduced.
- Enforces the gate server-side on every connect path, matching v16: the board
  card-menu "connect" postback, the board modal Connect click, and the detail
  block's lbConnect click. Hiding a button is no longer the only defense.
- Block setting descriptions updated to the v16 wording.
